Fixed misplaced chess pieces after scrolling in Chessboard page (during drag & drop) (#14)
The dynamic background was interfering with the Chessboard component. In fact, in the CSS class definitions of the background, there were positioning contexts that interfered with drag calculations. These have been removed, along with the fixed attachments that affected scroll behavior. Removed also any CSS transforms, as they could also interfere and are not needed for now.